backend/requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and utilities
python-dotenv==1.0.0
